+++ src/components/settings/SettingsPanels.tsx
@@ -1,5 +1,6 @@
 export { AiSettingsPanel } from "./AiSettingsPanel";
 export { LibrarySettingsPanel } from "./LibrarySettingsPanel";
 export { PublishingSettingsPanel } from "./PublishingSettingsPanel";
+export { ImageHostingSettingsPanel } from "./ImageHostingSettingsPanel";
 export { SettingsAboutPanel } from "./SettingsAboutPanel";
 export { WritingSettingsPanel } from "./WritingSettingsPanel";